test(hub): give hubcore and fspaths the hub's env redirect and EVENER_* scrub
cmd/evener-hub's TestMain redirected HOME, the XDG bases and CODEX_HOME into
a throwaway root and cleared every EVENER_* variable the product reads. None
of that reached the sub-package test binaries under cmd/evener-hub/internal/,
which had no TestMain at all: their isolation was per-test t.Setenv at the
handful of sites that touch a HOME/XDG-derived default. A test added beside
them that reaches such a default reads and writes the developer's real
~/.config/evener and ~/.local/state/evener, and a developer's shell
EVENER_PROVIDERS_CONFIG reaches every test that does not override it.

Extract the redirect, the scrub, the Go-cache pinning and the containment
check into cmd/evener-hub/internal/hubtestenv, and call it from a TestMain in
hubcore, in fspaths and in the hub package itself. The hub keeps what is its
own: the detach-helper early exit and the pre-run abort over its own default
resolvers.

Each sub-package gets one guard test asserting os.UserHomeDir and the three
XDG bases resolve under the throwaway root, mirroring the hub's.
